+++ packetbeat/protos/mongodb/mongodb.go
@@ -14,33 +14,33 @@ import (
 	"github.com/elastic/beats/packetbeat/protos/tcp"
 )
 
+var debugf = logp.MakeDebug("mongodb")
+
 type Mongodb struct {
 	// config
-	Ports          []int
-	Send_request   bool
-	Send_response  bool
-	Max_docs       int
-	Max_doc_length int
+	Ports        []int
+	SendRequest  bool
+	SendResponse bool
+	MaxDocs      int
+	MaxDocLength int
 
-	transactions       *common.Cache
+	requests           *common.Cache
+	responses          *common.Cache
 	transactionTimeout time.Duration
 
 	results publisher.Client
 }
 
-func (mongodb *Mongodb) getTransaction(k common.HashableTcpTuple) *MongodbTransaction {
-	v := mongodb.transactions.Get(k)
-	if v != nil {
-		return v.(*MongodbTransaction)
-	}
-	return nil
+type transactionKey struct {
+	tcp common.HashableTcpTuple
+	id  int
 }
 
 func (mongodb *Mongodb) InitDefaults() {
-	mongodb.Send_request = false
-	mongodb.Send_response = false
-	mongodb.Max_docs = 10
-	mongodb.Max_doc_length = 5000
+	mongodb.SendRequest = false
+	mongodb.SendResponse = false
+	mongodb.MaxDocs = 10
+	mongodb.MaxDocLength = 5000
 	mongodb.transactionTimeout = protos.DefaultTransactionExpiration
 }
 
@@ -48,16 +48,16 @@ func (mongodb *Mongodb) setFromConfig(config config.Mongodb) error {
 	mongodb.Ports = config.Ports
 
 	if config.SendRequest != nil {
-		mongodb.Send_request = *config.SendRequest
+		mongodb.SendRequest = *config.SendRequest
 	}
 	if config.SendResponse != nil {
-		mongodb.Send_response = *config.SendResponse
+		mongodb.SendResponse = *config.SendResponse
 	}
 	if config.Max_docs != nil {
-		mongodb.Max_docs = *config.Max_docs
+		mongodb.MaxDocs = *config.Max_docs
 	}
 	if config.Max_doc_length != nil {
-		mongodb.Max_doc_length = *config.Max_doc_length
+		mongodb.MaxDocLength = *config.Max_doc_length
 	}
 	if config.TransactionTimeout != nil && *config.TransactionTimeout > 0 {
 		mongodb.transactionTimeout = time.Duration(*config.TransactionTimeout) * time.Second
@@ -70,7 +70,7 @@ func (mongodb *Mongodb) GetPorts() []int {
 }
 
 func (mongodb *Mongodb) Init(test_mode bool, results publisher.Client) error {
-	logp.Debug("mongodb", "Init a MongoDB protocol parser")
+	debugf("Init a MongoDB protocol parser")
 
 	mongodb.InitDefaults()
 	if !test_mode {
@@ -80,10 +80,14 @@ func (mongodb *Mongodb) Init(test_mode bool, results publisher.Client) error {
 		}
 	}
 
-	mongodb.transactions = common.NewCache(
+	mongodb.requests = common.NewCache(
 		mongodb.transactionTimeout,
 		protos.DefaultTransactionHashSize)
-	mongodb.transactions.StartJanitor(mongodb.transactionTimeout)
+	mongodb.requests.StartJanitor(mongodb.transactionTimeout)
+	mongodb.responses = common.NewCache(
+		mongodb.transactionTimeout,
+		protos.DefaultTransactionHashSize)
+	mongodb.responses.StartJanitor(mongodb.transactionTimeout)
 	mongodb.results = results
 
 	return nil
@@ -93,162 +97,219 @@ func (mongodb *Mongodb) ConnectionTimeout() time.Duration {
 	return mongodb.transactionTimeout
 }
 
-func (mongodb *Mongodb) Parse(pkt *protos.Packet, tcptuple *common.TcpTuple, dir uint8,
-	private protos.ProtocolData) protos.ProtocolData {
-
-	logp.Debug("mongodb", "Parse method triggered")
-
+func (mongodb *Mongodb) Parse(
+	pkt *protos.Packet,
+	tcptuple *common.TcpTuple,
+	dir uint8,
+	private protos.ProtocolData,
+) protos.ProtocolData {
 	defer logp.Recover("ParseMongodb exception")
+	debugf("Parse method triggered")
 
-	// Either fetch or initialize current data struct for this parser
-	priv := mongodbPrivateData{}
-	if private != nil {
-		var ok bool
-		priv, ok = private.(mongodbPrivateData)
-		if !ok {
-			priv = mongodbPrivateData{}
-		}
+	conn := ensureMongodbConnection(private)
+	conn = mongodb.doParse(conn, pkt, tcptuple, dir)
+	if conn == nil {
+		return nil
+	}
+	return conn
+}
+
+func ensureMongodbConnection(private protos.ProtocolData) *mongodbConnectionData {
+	if private == nil {
+		return &mongodbConnectionData{}
 	}
 
-	if priv.Data[dir] == nil {
-		priv.Data[dir] = &MongodbStream{
-			tcptuple: tcptuple,
-			data:     pkt.Payload,
-			message:  &MongodbMessage{Ts: pkt.Ts},
-		}
-	} else {
-		// concatenate bytes
-		priv.Data[dir].data = append(priv.Data[dir].data, pkt.Payload...)
-		if len(priv.Data[dir].data) > tcp.TCP_MAX_DATA_IN_STREAM {
-			logp.Debug("mongodb", "Stream data too large, dropping TCP stream")
-			priv.Data[dir] = nil
-			return priv
-		}
+	priv, ok := private.(*mongodbConnectionData)
+	if !ok {
+		logp.Warn("mongodb connection data type error, create new one")
+		return &mongodbConnectionData{}
 	}
-
-	stream := priv.Data[dir]
-	for len(stream.data) > 0 {
-		if stream.message == nil {
-			stream.message = &MongodbMessage{Ts: pkt.Ts}
-		}
-
-		ok, complete := mongodbMessageParser(priv.Data[dir])
-
-		if !ok {
-			// drop this tcp stream. Will retry parsing with the next
-			// segment in it
-			priv.Data[dir] = nil
-			logp.Debug("mongodb", "Ignore Mongodb message. Drop tcp stream. Try parsing with the next segment")
-			return priv
-		}
-
-		if complete {
-
-			logp.Debug("mongodb", "MongoDB message complete")
-
-			// all ok, go to next level
-			mongodb.handleMongodb(stream.message, tcptuple, dir)
-
-			// and reset message
-			stream.PrepareForNewMessage()
-		} else {
-			// wait for more data
-			logp.Debug("mongodb", "MongoDB wait for more data before parsing message")
-			break
-		}
+	if priv == nil {
+		logp.Warn("Unexpected: mongodb connection data not set, create new one")
+		return &mongodbConnectionData{}
 	}
 
 	return priv
 }
 
-func (mongodb *Mongodb) handleMongodb(m *MongodbMessage, tcptuple *common.TcpTuple,
-	dir uint8) {
+func (mongodb *Mongodb) doParse(
+	conn *mongodbConnectionData,
+	pkt *protos.Packet,
+	tcptuple *common.TcpTuple,
+	dir uint8,
+) *mongodbConnectionData {
+	st := conn.Streams[dir]
+	if st == nil {
+		st = newStream(pkt, tcptuple)
+		conn.Streams[dir] = st
+		debugf("new stream: %p (dir=%v, len=%v)", st, dir, len(pkt.Payload))
+	} else {
+		// concatenate bytes
+		st.data = append(st.data, pkt.Payload...)
+		if len(st.data) > tcp.TCP_MAX_DATA_IN_STREAM {
+			debugf("Stream data too large, dropping TCP stream")
+			conn.Streams[dir] = nil
+			return conn
+		}
+	}
+
+	for len(st.data) > 0 {
+		if st.message == nil {
+			st.message = &mongodbMessage{Ts: pkt.Ts}
+		}
+
+		ok, complete := mongodbMessageParser(st)
+		if !ok {
+			// drop this tcp stream. Will retry parsing with the next
+			// segment in it
+			conn.Streams[dir] = nil
+			debugf("Ignore Mongodb message. Drop tcp stream. Try parsing with the next segment")
+			return conn
+		}
+
+		if !complete {
+			// wait for more data
+			debugf("MongoDB wait for more data before parsing message")
+			break
+		}
+
+		// all ok, go to next level and reset stream for new message
+		debugf("MongoDB message complete")
+		mongodb.handleMongodb(conn, st.message, tcptuple, dir)
+		st.PrepareForNewMessage()
+	}
+
+	return conn
+}
+
+func newStream(pkt *protos.Packet, tcptuple *common.TcpTuple) *stream {
+	s := &stream{
+		tcptuple: tcptuple,
+		data:     pkt.Payload,
+		message:  &mongodbMessage{Ts: pkt.Ts},
+	}
+	return s
+}
+
+func (mongodb *Mongodb) handleMongodb(
+	conn *mongodbConnectionData,
+	m *mongodbMessage,
+	tcptuple *common.TcpTuple,
+	dir uint8,
+) {
 
 	m.TcpTuple = *tcptuple
 	m.Direction = dir
 	m.CmdlineTuple = procs.ProcWatcher.FindProcessesTuple(tcptuple.IpPort())
 
 	if m.IsResponse {
-		logp.Debug("mongodb", "MongoDB response message")
-		mongodb.receivedMongodbResponse(m)
+		debugf("MongoDB response message")
+		mongodb.onResponse(conn, m)
 	} else {
-		logp.Debug("mongodb", "MongoDB request message")
-		mongodb.receivedMongodbRequest(m)
+		debugf("MongoDB request message")
+		mongodb.onRequest(conn, m)
 	}
 }
 
-func (mongodb *Mongodb) receivedMongodbRequest(msg *MongodbMessage) {
-	// Add it to the HT
-	tuple := msg.TcpTuple
-
-	trans := mongodb.getTransaction(tuple.Hashable())
-	if trans != nil {
-		if trans.Mongodb != nil {
-			logp.Warn("Two requests without a Response. Dropping old request")
-		}
-	} else {
-		logp.Debug("mongodb", "Initialize new transaction from request")
-		trans = &MongodbTransaction{Type: "mongodb", tuple: tuple}
-		mongodb.transactions.Put(tuple.Hashable(), trans)
+func (mongodb *Mongodb) onRequest(conn *mongodbConnectionData, msg *mongodbMessage) {
+	// publish request only transaction
+	if !awaitsReply(msg.opCode) {
+		mongodb.onTransComplete(msg, nil)
+		return
 	}
 
-	trans.Mongodb = common.MapStr{}
+	id := msg.requestId
+	key := transactionKey{tcp: msg.TcpTuple.Hashable(), id: id}
 
-	trans.event = msg.event
+	// try to find matching response potentially inserted before
+	if v := mongodb.responses.Delete(key); v != nil {
+		resp := v.(*mongodbMessage)
+		mongodb.onTransComplete(msg, resp)
+		return
+	}
 
-	trans.method = msg.method
-
-	trans.cmdline = msg.CmdlineTuple
-	trans.ts = msg.Ts
-	trans.Ts = int64(trans.ts.UnixNano() / 1000) // transactions have microseconds resolution
-	trans.JsTs = msg.Ts
-	trans.Src = common.Endpoint{
-		Ip:   msg.TcpTuple.Src_ip.String(),
-		Port: msg.TcpTuple.Src_port,
-		Proc: string(msg.CmdlineTuple.Src),
+	// insert into cache for correlation
+	old := mongodb.requests.Put(key, msg)
+	if old != nil {
+		logp.Warn("Two requests without a Response. Dropping old request")
 	}
-	trans.Dst = common.Endpoint{
-		Ip:   msg.TcpTuple.Dst_ip.String(),
-		Port: msg.TcpTuple.Dst_port,
-		Proc: string(msg.CmdlineTuple.Dst),
-	}
-	if msg.Direction == tcp.TcpDirectionReverse {
-		trans.Src, trans.Dst = trans.Dst, trans.Src
-	}
-	trans.params = msg.params
-	trans.resource = msg.resource
-	trans.BytesIn = msg.messageLength
 }
 
-func (mongodb *Mongodb) receivedMongodbResponse(msg *MongodbMessage) {
+func (mongodb *Mongodb) onResponse(conn *mongodbConnectionData, msg *mongodbMessage) {
+	id := msg.responseTo
+	key := transactionKey{tcp: msg.TcpTuple.Hashable(), id: id}
 
-	trans := mongodb.getTransaction(msg.TcpTuple.Hashable())
-	if trans == nil {
-		logp.Warn("Response from unknown transaction. Ignoring.")
+	// try to find matching request
+	if v := mongodb.requests.Delete(key); v != nil {
+		requ := v.(*mongodbMessage)
+		mongodb.onTransComplete(requ, msg)
 		return
 	}
-	// check if the request was received
-	if trans.Mongodb == nil {
-		logp.Warn("Response from unknown transaction. Ignoring.")
-		return
 
-	}
+	// insert into cache for correlation
+	mongodb.responses.Put(key, msg)
+}
 
-	// Merge request and response events attributes
-	for k, v := range msg.event {
-		trans.event[k] = v
-	}
-
-	trans.error = msg.error
-	trans.documents = msg.documents
-
-	trans.ResponseTime = int32(msg.Ts.Sub(trans.ts).Nanoseconds() / 1e6) // resp_time in milliseconds
-	trans.BytesOut = msg.messageLength
+func (mongodb *Mongodb) onTransComplete(requ, resp *mongodbMessage) {
+	trans := newTransaction(requ, resp)
+	debugf("Mongodb transaction completed: %s", trans.Mongodb)
 
 	mongodb.publishTransaction(trans)
-	mongodb.transactions.Delete(trans.tuple.Hashable())
+}
 
-	logp.Debug("mongodb", "Mongodb transaction completed: %s", trans.Mongodb)
+func newTransaction(requ, resp *mongodbMessage) *transaction {
+	trans := &transaction{Type: "mongodb"}
+
+	// fill request
+	if requ != nil {
+		trans.tuple = requ.TcpTuple
+
+		trans.Mongodb = common.MapStr{}
+		trans.event = requ.event
+		trans.method = requ.method
+
+		trans.cmdline = requ.CmdlineTuple
+		trans.ts = requ.Ts
+		trans.Ts = int64(trans.ts.UnixNano() / 1000) // transactions have microseconds resolution
+		trans.JsTs = requ.Ts
+		trans.Src = common.Endpoint{
+			Ip:   requ.TcpTuple.Src_ip.String(),
+			Port: requ.TcpTuple.Src_port,
+			Proc: string(requ.CmdlineTuple.Src),
+		}
+		trans.Dst = common.Endpoint{
+			Ip:   requ.TcpTuple.Dst_ip.String(),
+			Port: requ.TcpTuple.Dst_port,
+			Proc: string(requ.CmdlineTuple.Dst),
+		}
+		if requ.Direction == tcp.TcpDirectionReverse {
+			trans.Src, trans.Dst = trans.Dst, trans.Src
+		}
+		trans.params = requ.params
+		trans.resource = requ.resource
+		trans.BytesIn = requ.messageLength
+	}
+
+	// fill response
+	if resp != nil {
+		if requ == nil {
+			// TODO: reverse tuple?
+			trans.tuple = resp.TcpTuple
+		}
+
+		for k, v := range resp.event {
+			trans.event[k] = v
+		}
+
+		trans.error = resp.error
+		trans.documents = resp.documents
+
+		trans.ResponseTime = int32(resp.Ts.Sub(trans.ts).Nanoseconds() / 1e6) // resp_time in milliseconds
+		trans.BytesOut = resp.messageLength
+
+	}
+
+	return trans
 }
 
 func (mongodb *Mongodb) GapInStream(tcptuple *common.TcpTuple, dir uint8,
@@ -276,7 +337,7 @@ func copy_map_without_key(d map[string]interface{}, key string) map[string]inter
 	return res
 }
 
-func reconstructQuery(t *MongodbTransaction, full bool) (query string) {
+func reconstructQuery(t *transaction, full bool) (query string) {
 	query = t.resource + "." + t.method + "("
 	if len(t.params) > 0 {
 		var err error
@@ -295,7 +356,7 @@ func reconstructQuery(t *MongodbTransaction, full bool) (query string) {
 			params, err = doc2str(t.params)
 		}
 		if err != nil {
-			logp.Debug("mongodb", "Error marshaling params: %v", err)
+			debugf("Error marshaling params: %v", err)
 		} else {
 			query += params
 		}
@@ -313,10 +374,10 @@ func reconstructQuery(t *MongodbTransaction, full bool) (query string) {
 	return
 }
 
-func (mongodb *Mongodb) publishTransaction(t *MongodbTransaction) {
+func (mongodb *Mongodb) publishTransaction(t *transaction) {
 
 	if mongodb.results == nil {
-		logp.Debug("mongodb", "Try to publish transaction with null results")
+		debugf("Try to publish transaction with null results")
 		return
 	}
 
@@ -339,15 +400,15 @@ func (mongodb *Mongodb) publishTransaction(t *MongodbTransaction) {
 	event["src"] = &t.Src
 	event["dst"] = &t.Dst
 
-	if mongodb.Send_request {
+	if mongodb.SendRequest {
 		event["request"] = reconstructQuery(t, true)
 	}
-	if mongodb.Send_response {
+	if mongodb.SendResponse {
 		if len(t.documents) > 0 {
 			// response field needs to be a string
 			docs := make([]string, 0, len(t.documents))
 			for i, doc := range t.documents {
-				if mongodb.Max_docs > 0 && i >= mongodb.Max_docs {
+				if mongodb.MaxDocs > 0 && i >= mongodb.MaxDocs {
 					docs = append(docs, "[...]")
 					break
 				}
@@ -355,8 +416,8 @@ func (mongodb *Mongodb) publishTransaction(t *MongodbTransaction) {
 				if err != nil {
 					logp.Warn("Failed to JSON marshal document from Mongo: %v (error: %v)", doc, err)
 				} else {
-					if mongodb.Max_doc_length > 0 && len(str) > mongodb.Max_doc_length {
-						str = str[:mongodb.Max_doc_length] + " ..."
+					if mongodb.MaxDocLength > 0 && len(str) > mongodb.MaxDocLength {
+						str = str[:mongodb.MaxDocLength] + " ..."
 					}
 					docs = append(docs, str)
 				}